requirements.txt
absl-py==2.0.0
accelerate==0.24.1
aiohttp==3.8.6
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
clip-interrogator==0.6.0
contourpy==1.1.1
cycler==0.12.1
datasets==2.14.6
dill==0.3.7
easydict==1.11
einops==0.7.0
filelock==3.12.4
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.1
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
keras==2.14.0
kiwisolver==1.4.5
kornia==0.7.0
libclang==16.0.6
lightning-bolts==0.7.0
lightning-lite==1.8.0
lightning-utilities==0.9.0
llvmlite==0.41.1
lpips==0.1.4
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2
nltk==3.8.1
numba==0.58.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
open-clip-torch==2.23.0
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
paintmind @ git+https://github.com/Qiyuan-Ge/PaintMind.git@733bcf4e4e44dda1f9b959beb9fd96b13de2dfb1
pandas==2.1.2
pilgram==1.2.1
Pillow==10.1.0
piqa==1.3.2
protobuf==4.24.4
psutil==5.9.6
pyarrow==14.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==1.9.5
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
scipy==1.11.3
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
timm==0.9.8
tokenizers==0.14.1
torch==2.1.0
torchmetrics==0.11.4
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.9
Werkzeug==3.0.1
wrapt==1.14.1
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
